requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
Django==4.0.2
djangorestframework==3.13.1
drf-spectacular==0.21.2
frozenlist==1.3.0
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
jsonschema==4.4.0
lxml==4.7.1
multidict==6.0.2
pycares==4.1.2
pycparser==2.21
pyee==8.2.2
pyppeteer==1.0.2
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
zipp==3.7.0
